+++ jaguar_serializer_cli/CHANGELOG.md
@@ -1,8 +1,16 @@
 # Changelog
 
-## 2.1.15
+## 2.1.17
 
 + Fix for dart-dev64.1
+
+## 2.1.16
+
++ dynamic field
+
+## 2.1.15
+
++ Bug fix: [Final non-nullable contructor](https://github.com/Jaguar-dart/jaguar_serializer/issues/144)
 
 ## 2.1.14